chore: rename patient dashboard files to better names
